Add videos sort. Save sort preferences with cookies.
